refactor: extract more from src/markdown/transform-markdown.ts
