+++ src/markdown/get-input-asts.ts
@@ -1,0 +1,13 @@
+import { Nodes } from "npm:@types/mdast";
+import { markdownToAst } from "../ast/markdown-to-ast.ts";
+
+export function getInputAsts(
+  inputs: Record<string, string>,
+): Record<string, Nodes> {
+  return Object.fromEntries(
+    Object.entries(inputs).map(([inputPath, input]) => [
+      inputPath,
+      markdownToAst(input),
+    ]),
+  );
+}
