+++ test/traverse-cases.ts
@@ -1,7 +1,7 @@
 import { walk, WalkEntry } from "std/fs/walk.ts";
 import { Nodes } from "npm:@types/mdast";
 import { markdownToAst } from "../src/ast/markdown-to-ast.ts";
-import { DeleteOrWriteFile } from "../src/markdown/transform-markdown.ts";
+import { DeleteOrWriteFile } from "../src/commands/output-command.ts";
 
 /**
  * Corresponds to a directory with test cases in it.
